Рефакторинг открытия модалок
