+++ src/entities/record/ui/DeleteDialog/DeleteDialog.test.tsx
