+++ src/entities/record/ui/DeleteDialog/DeleteDialog.tsx
