+++ src/entities/record/ui/DeleteDialog/index.ts
@@ -1,0 +1,1 @@
+export { DeleteDialog } from './DeleteDialog';
