--- src/features/records/ui/Records/DeleteDialog/index.ts
@@ -1,3 +1,0 @@
-export { DeleteDialog } from './DeleteDialog';
-
-export { useDeleteDialog } from './useDeleteDialog';
