+++ src/pages/HomePage/ui/DayModal/DayModal.css.ts
