+++ src/pages/HomePage/ui/DayModal/index.ts
@@ -1,0 +1,1 @@
+export { DayModal } from './DayModal';
