--- src/pages/HomePage/ui/Modal/index.ts
@@ -1,1 +1,0 @@
-export { Modal } from './Modal';
