+++ src/shared/appSettings.ts
@@ -1,6 +1,4 @@
 export const AppSettings = {
   DOSES: ['10', '20', '30'],
   DEFAULT_DOSE: '30',
-  DAY_TARGET: '20',
-  TARGET_DOSE: 10500,
 } as const;
