--- src/shared/assets/icons/PlusIcon/index.ts
@@ -1,1 +1,0 @@
-export { PlusIcon } from './PlusIcon';
