+++ src/shared/providers/RecordsProvider/lib/DBController/index.ts
@@ -1,0 +1,1 @@
+export { DBController, type DBRecord } from './DBController';
