--- src/shared/types/Record.ts
@@ -1,6 +1,0 @@
-export type RecordType = {
-  id?: number;
-  datetime: string;
-  dose: string;
-  targetDose: string;
-};
