--- src/shared/ui/AddButton/index.ts
@@ -1,1 +1,0 @@
-export { AddButton } from './AddButton';
